Fix syntax for string literal in square brackets (#2)
Since https://docs.scala-lang.org/sips/42.type.html which is implemented
in scala 2.13 and in scala 3 it possible to use string literals as
singleton types. So code like
```
someFunc["abc"]
```
is valid. Currently this code is not hightlighted correctly and worse if
there is an unclosed `(` in the string it breaks the formating in the
rest of the file.